backend/metadata_processor/requirements/requirements.txt
pika==1.3.2
PyYAML==6.0.1
pillow==10.3.0
numpy==1.26.4
opencv-python==4.8.0.74